refactor: remove node tooling
