+++ README.md
@@ -1,19 +1,9 @@
 # Сердца на грани чудес
 
-Прототип визуальной новеллы для Яндекс Игр. Проект реализуется на TypeScript с целью демонстрации движка, загружающего эпизоды из JSON.
+Прототип визуальной новеллы для Яндекс Игр. Теперь проект реализован на чистом HTML5 и JavaScript без использования Node.js.
 
 Полное техническое задание находится в [docs/TECHNICAL_SPEC.md](docs/TECHNICAL_SPEC.md).
 
-## Разработка
+## Запуск
 
-```bash
-npm install
-npm test
-npm run build
-```
-
-`npm test` запускает проверку TypeScript-компиляции.
-
-`npm run build` компилирует исходники в папку `public`. Для корректной работы `public/index.html` необходимо раздавать по HTTP (например, `npx http-server public`), так как `fetch` не может загружать данные по схеме `file://`.
-
-На текущем этапе игра выводит сообщения в консоль и не содержит пользовательского интерфейса.
+Все файлы игры находятся в папке `public`. Для корректной работы `public/index.html` необходимо раздавать по HTTP (например, `python3 -m http.server 8000 --directory public`), так как `fetch` не может загружать данные по схеме `file://`.
