--- package.json
@@ -1,12 +1,0 @@
-{
-  "name": "game-miracles-hearts",
-  "version": "0.1.0",
-  "type": "module",
-  "scripts": {
-    "test": "tsc --noEmit",
-    "build": "tsc && cp -r content public/content && cp -r assets public/assets"
-  },
-  "devDependencies": {
-    "typescript": "^5.3.3"
-  }
-}
